pyproject.toml
[project]
name = "home-service-flyio"
version = "0.1.0"
description = ""
authors = [
    {name = "JS",email = "dev@voidst.one"}
]
readme = "README.md"
requires-python = ">=3.14,<4.0"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "django-allauth (>=65.14.0,<66.0.0)",
    "django-bootstrap5 (>=26.1,<27.0)",
    "django-crispy-forms (>=2.5,<3.0)",
    "crispy-bootstrap5 (>=2025.6,<2026.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "haversine (>=2.9.0,<3.0.0)",
    "django-auditlog (>=3.4.1,<4.0.0)",
    "django-tailwind-cli (>=4.5.1,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
